requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
scikit-learn==1.5.1
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2
gunicorn==22.0.0
pytest==8.2.2
python-dotenv==1.0.1
